Formatting fixes

+++ code/FuseVelPosNED.cpp
@@ -20,22 +20,22 @@
 
 // run one state and covariance update step using a combination of NED Vel and Pos measurements
 void FuseVelPosNED(
-        float nextStates[24], // state output
-        float nextP[24][24],  // covariance output
-        float innovation[6], // innovation output
-        float varInnov[6], // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        float accNavMag, // magnitude of navigation accel (- used to adjust GPS obs variance (m/s^2)
-        bool FuseGPSData, // this boolean causes the PosNE and VelNED obs to be fused
-        float VelNED[3], // North, East, Down velocity obs (m/s)
-        bool useVelD, // this boolean casues the D component of the VelNED vector to be used
-        float PosNE[3], // North, East position obs (m)
-        float StatesAtGpsTime[24], // States at the effective measurement time for PosNE and VelNED measurements
-        bool FuseHgtData, // this boolean causes the HgtMea obs to be fused
-        float HgtMea, //  measured height (m)
-        float StatesAtHgtTime[24], // States at the effective measurement time for the HgtMea measurement
-        bool useAirspeed) // this boolean indicates that airspeed measurements are also being used
+    float nextStates[24], // state output
+    float nextP[24][24],  // covariance output
+    float innovation[6], // innovation output
+    float varInnov[6], // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    float accNavMag, // magnitude of navigation accel (- used to adjust GPS obs variance (m/s^2)
+    bool FuseGPSData, // this boolean causes the PosNE and VelNED obs to be fused
+    float VelNED[3], // North, East, Down velocity obs (m/s)
+    bool useVelD, // this boolean casues the D component of the VelNED vector to be used
+    float PosNE[3], // North, East position obs (m)
+    float StatesAtGpsTime[24], // States at the effective measurement time for PosNE and VelNED measurements
+    bool FuseHgtData, // this boolean causes the HgtMea obs to be fused
+    float HgtMea, //  measured height (m)
+    float StatesAtHgtTime[24], // States at the effective measurement time for the HgtMea measurement
+    bool useAirspeed) // this boolean indicates that airspeed measurements are also being used
 {
 
 // declare variables used by fault isolation logic
@@ -60,11 +60,11 @@ void FuseVelPosNED(
     float hgtInnov = 0.0;
 
 // declare indices used to access arrays
-    int stateIndex;
-    int obsIndex;
-    int i;
-    int j;
-    int iMax;
+    unsigned short int stateIndex;
+    unsigned short int obsIndex;
+    unsigned short int i;
+    unsigned short int j;
+    unsigned short int iMax;
 
 // declare variables used by state and covariance update calculations
     float velErr;
@@ -126,12 +126,13 @@ void FuseVelPosNED(
         // Set innovation variances to zero default
         for (i = 0; i<=5; i++)
         {
-             varInnov[i] = 0.0;
+            varInnov[i] = 0.0;
         }
         // calculate innovations and check GPS data validity against limits using a 5-sigma test
         if (FuseGPSData)
         {
-            if (useVelD) iMax = 2; else iMax = 1;
+            if (useVelD) iMax = 2;
+            else iMax = 1;
             for (i = 0; i<=iMax; i++)
             {
                 velInnov[i] = StatesAtGpsTime[i+4] - VelNED[i];
