+++ code/main.cpp
@@ -13,46 +13,46 @@ void CovariancePrediction(
     bool useCompass);
 
 void FuseVelPosNED(
-        float nextStates[24], // state output
-        float nextP[24][24],  // covariance output
-        float innovation[6], // innovation output
-        float varInnov[6], // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        float accNavMag, // magnitude of navigation accel (- used to adjust GPS obs variance (m/s^2)
-        bool FuseGPSData, // this boolean causes the PosNE and VelNED obs to be fused
-        float VelNED[3], // North, East, Down velocity obs (m/s)
-        bool useVelD, // this boolean casues the D component of the VelNED vector to be used
-        float PosNE[3], // North, East position obs (m)
-        float StatesAtGpsTime[24], // States at the effective measurement time for PosNE and VelNED measurements
-        bool FuseHgtData, // this boolean causes the HgtMea obs to be fused
-        float HgtMea, //  measured height (m)
-        float StatesAtHgtTime[24], // States at the effective measurement time for the HgtMea measurement
-        bool useAirspeed); // this boolean indicates that airspeed measurements are also being used
+    float nextStates[24], // state output
+    float nextP[24][24],  // covariance output
+    float innovation[6], // innovation output
+    float varInnov[6], // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    float accNavMag, // magnitude of navigation accel (- used to adjust GPS obs variance (m/s^2)
+    bool FuseGPSData, // this boolean causes the PosNE and VelNED obs to be fused
+    float VelNED[3], // North, East, Down velocity obs (m/s)
+    bool useVelD, // this boolean casues the D component of the VelNED vector to be used
+    float PosNE[3], // North, East position obs (m)
+    float StatesAtGpsTime[24], // States at the effective measurement time for PosNE and VelNED measurements
+    bool FuseHgtData, // this boolean causes the HgtMea obs to be fused
+    float HgtMea, //  measured height (m)
+    float StatesAtHgtTime[24], // States at the effective measurement time for the HgtMea measurement
+    bool useAirspeed); // this boolean indicates that airspeed measurements are also being used
 
 void FuseMagnetometer(
-        float nextStates[24], // state output
-        float nextP[24][24], // covariance output
-        float innovation[6], // innovation output
-        float varInnov[6], // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        bool FuseData, // boolean true when magnetometer data is to be fused
-        float MagData[3], // magnetometer flux radings in X,Y,Z body axes
-        float StatesAtMeasTime[24], // filter satates at the effective measurement time
-        bool useCompass); // boolean true if magnetometer data is being used
+    float nextStates[24], // state output
+    float nextP[24][24], // covariance output
+    float innovation[6], // innovation output
+    float varInnov[6], // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    bool FuseData, // boolean true when magnetometer data is to be fused
+    float MagData[3], // magnetometer flux radings in X,Y,Z body axes
+    float StatesAtMeasTime[24], // filter satates at the effective measurement time
+    bool useCompass); // boolean true if magnetometer data is being used
 
 void FuseAirspeed(
-        float nextStates[24], // state output
-        float nextP[24][24], // covariance output
-        float innovation, // innovation output
-        float varInnov, // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        bool FuseData, // boolean true when airspeed data is to be fused
-        float VtasMeas, // true airspeed measurement (m/s)
-        float StatesAtMeasTime[24], // filter states at the effective measurement time
-        bool useAirspeed); // boolean true if airspeed data is being used
+    float nextStates[24], // state output
+    float nextP[24][24], // covariance output
+    float innovation, // innovation output
+    float varInnov, // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    bool FuseData, // boolean true when airspeed data is to be fused
+    float VtasMeas, // true airspeed measurement (m/s)
+    float StatesAtMeasTime[24], // filter states at the effective measurement time
+    bool useAirspeed); // boolean true if airspeed data is being used
 
 void zeroRows(float covMat[24][24], unsigned short int first, unsigned short int last);
 
@@ -853,22 +853,22 @@ void CovariancePrediction(
 }
 
 void FuseVelPosNED(
-        float nextStates[24], // state output
-        float nextP[24][24],  // covariance output
-        float innovation[6], // innovation output
-        float varInnov[6], // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        float accNavMag, // magnitude of navigation accel (- used to adjust GPS obs variance (m/s^2)
-        bool FuseGPSData, // this boolean causes the PosNE and VelNED obs to be fused
-        float VelNED[3], // North, East, Down velocity obs (m/s)
-        bool useVelD, // this boolean casues the D component of the VelNED vector to be used
-        float PosNE[3], // North, East position obs (m)
-        float StatesAtGpsTime[24], // States at the effective measurement time for PosNE and VelNED measurements
-        bool FuseHgtData, // this boolean causes the HgtMea obs to be fused
-        float HgtMea, //  measured height (m)
-        float StatesAtHgtTime[24], // States at the effective measurement time for the HgtMea measurement
-        bool useAirspeed) // this boolean indicates that airspeed measurements are also being used
+    float nextStates[24], // state output
+    float nextP[24][24],  // covariance output
+    float innovation[6], // innovation output
+    float varInnov[6], // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    float accNavMag, // magnitude of navigation accel (- used to adjust GPS obs variance (m/s^2)
+    bool FuseGPSData, // this boolean causes the PosNE and VelNED obs to be fused
+    float VelNED[3], // North, East, Down velocity obs (m/s)
+    bool useVelD, // this boolean casues the D component of the VelNED vector to be used
+    float PosNE[3], // North, East position obs (m)
+    float StatesAtGpsTime[24], // States at the effective measurement time for PosNE and VelNED measurements
+    bool FuseHgtData, // this boolean causes the HgtMea obs to be fused
+    float HgtMea, //  measured height (m)
+    float StatesAtHgtTime[24], // States at the effective measurement time for the HgtMea measurement
+    bool useAirspeed) // this boolean indicates that airspeed measurements are also being used
 {
 
 // declare variables used by fault isolation logic
@@ -893,11 +893,11 @@ void FuseVelPosNED(
     float hgtInnov = 0.0;
 
 // declare indices used to access arrays
-    int stateIndex;
-    int obsIndex;
-    int i;
-    int j;
-    int iMax;
+    unsigned short int stateIndex;
+    unsigned short int obsIndex;
+    unsigned short int i;
+    unsigned short int j;
+    unsigned short int iMax;
 
 // declare variables used by state and covariance update calculations
     float velErr;
@@ -908,8 +908,8 @@ void FuseVelPosNED(
     float SK;
     float K[24];
     float quatMag;
-    int startIndex;
-    int endIndex;
+    unsigned short int startIndex;
+    unsigned short int endIndex;
 
 //Default action - pass through states and covariances
     for (i=0; i<=23; i++)
@@ -959,12 +959,13 @@ void FuseVelPosNED(
         // Set innovation variances to zero default
         for (i = 0; i<=5; i++)
         {
-             varInnov[i] = 0.0;
+            varInnov[i] = 0.0;
         }
         // calculate innovations and check GPS data validity against limits using a 5-sigma test
         if (FuseGPSData)
         {
-            if (useVelD) iMax = 2; else iMax = 1;
+            if (useVelD) iMax = 2;
+            else iMax = 1;
             for (i = 0; i<=iMax; i++)
             {
                 velInnov[i] = StatesAtGpsTime[i+4] - VelNED[i];
@@ -1108,16 +1109,16 @@ void FuseVelPosNED(
 }
 
 void FuseMagnetometer(
-        float nextStates[24], // state output
-        float nextP[24][24], // covariance output
-        float innovation[3], // innovation output
-        float varInnov[3], // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        bool FuseData, // boolean true when magnetometer data is to be fused
-        float MagData[3], // magnetometer flux radings in X,Y,Z body axes
-        float StatesAtMeasTime[24], // filter satates at the effective measurement time
-        bool useCompass) // boolean true if magnetometer data is being used
+    float nextStates[24], // state output
+    float nextP[24][24], // covariance output
+    float innovation[3], // innovation output
+    float varInnov[3], // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    bool FuseData, // boolean true when magnetometer data is to be fused
+    float MagData[3], // magnetometer flux radings in X,Y,Z body axes
+    float StatesAtMeasTime[24], // filter satates at the effective measurement time
+    bool useCompass) // boolean true if magnetometer data is being used
 {
 
     static float q0 = 1.0;
@@ -1132,7 +1133,8 @@ void FuseMagnetometer(
     static float magZbias = 0.0;
     static unsigned short int obsIndex = 0;
     static float SH_MAG[9] = {0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0};
-    static float Tnb[3][3] = {
+    static float Tnb[3][3] =
+    {
         {1.0,0.0,0.0} ,
         {0.0,1.0,0.0} ,
         {0.0,0.0,1.0}
@@ -1408,16 +1410,16 @@ void FuseMagnetometer(
 }
 
 void FuseAirspeed(
-        float nextStates[24], // state output
-        float nextP[24][24], // covariance output
-        float innovation, // innovation output
-        float varInnov, // innovation variance output
-        float states[24], // state input
-        float P[24][24], // covariance input
-        bool FuseData, // boolean true when airspeed data is to be fused
-        float VtasMeas, // true airspeed measurement (m/s)
-        float StatesAtMeasTime[24], // filter states at the effective measurement time
-        bool useAirspeed) // boolean true if airspeed data is being used
+    float nextStates[24], // state output
+    float nextP[24][24], // covariance output
+    float innovation, // innovation output
+    float varInnov, // innovation variance output
+    float states[24], // state input
+    float P[24][24], // covariance input
+    bool FuseData, // boolean true when airspeed data is to be fused
+    float VtasMeas, // true airspeed measurement (m/s)
+    float StatesAtMeasTime[24], // filter states at the effective measurement time
+    bool useAirspeed) // boolean true if airspeed data is being used
 {
     float vn;
     float ve;
